Add the rest of the functions
Linked the playerChoice and compChoice functions together and made a loop that
will let you play indefinitely. Basic functionality is present, but future
extra things to add would be an option to set your preferred number of rounds,
your preferred points to win, or extra text/visuals. For now, this game is considered
done.